fix: Move out fetch items to needed
